+++ backend/app/controllers/pacientes_controller.rb
@@ -372,7 +372,8 @@ class PacientesController < ApplicationController
 
   def paciente_params
     allowed = %i[nombre apellido dni fecha_nacimiento es_paciente email telefono reprocann_numero reprocann_vencimiento reprocann_estado
-                 domicilio_calle domicilio_altura domicilio_piso domicilio_depto domicilio_barrio domicilio_ciudad]
+                 domicilio_calle domicilio_altura domicilio_piso domicilio_depto domicilio_barrio domicilio_ciudad
+                 envio_calle envio_altura envio_piso envio_depto envio_barrio envio_ciudad]
     if current_user&.admin? || current_user&.super_admin?
       allowed += %i[limite_dispensacion_mensual_g descuento_porcentaje]
     end